requirements.txt
fastapi == 0.68.1
uvicorn == 0.15.0
python-dotenv == 0.19.0